requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
livekit-plugins-google
livekit-plugins-noise-cancellation
requests
python-dotenv